Explain limitation of folding in the documentation
As the foldmethod=manual is being because syntax method is too slow, the folds
can be wrong if sessions are used and the file changed while Vim was closed,
and it may cause errors if the file shrinks and the session attempt to create
fold on inexistent lines.

Thus it would be better to remove the folds before creating sessions. Some
attempts were made by including a window local variable on the log windows
that created folds and them deleting the folds before the session was created:

    function! LogNav#EnableFold()
      " ...
      let w:lognav_fold_autocmd = 1
    endfunction

    " Autocmds {{{1
    augroup LogNav
      autocmd!
      au QuitPre * call LogNavDisableAllFolds()
    augroup END

    function! LogNavDisableAllFolds()
       let initTab = tabpagenr()
       let initWin = winnr()
       for tab in range(1, tabpagenr('$'))
          for win in range(1, tabpagewinnr(tab, '$'))
             if gettabwinvar(tab, win, 'lognav_fold_autocmd', 0)
                exe tab . "tabn"
                exe win . "wincmd w"
                normal! zE
                redraw!
             endif
          endfor
       endfor
       exe initTab . "tabn"
       exe initWin . "wincmd w"
    endfunction

Some other autocmds were tried (e.g.: <buffer> BufWinLeave, QuitPre), and
other methods for deleting the folds, such as changing 'fdm' to 'expr', but
none worked. So it is better to acept the current form until the
foldmethod=syntax is fixed on Vim.